+++ packages/spotlight/SpotlightRootDecorator/SpotlightRootDecorator.js
@@ -35,9 +35,10 @@ const defaultConfig = {
  * Constructs a Higher-order Component that initializes and enables Spotlight
  * 5-way navigation within an application.
  *
- * @example
+ * Example:
+ * ```
  *	const App = SpotlightRootDecorator(ApplicationView);
- *
+ * ```
  * @param  {Object} defaultConfig Set of default configuration parameters
  * @param  {Function} Higher-order component
  *